Add a first cut/outline explainer
This is very very rough, it needs a lot of help, but it is a start. Typically explainers are .md and just point to a gh repo hosted explainer. We could do otherwise, but editing is very easy that way. Unfortunately we've got to merge this before I think we can easily link it up, so I'm not deleting the other one for now just yet, we can do that post merge.